requirements.txt
aiofiles==23.1.0
aiogram==2.25.1
aiohttp==3.8.5
folium==0.14.0
google-api-core==2.11.1
google-api-python-client==2.48.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
gspread==5.7.2
gspread-asyncio==1.8.1
pandas
PyPDF2==1.27.12
PyPDF3==1.0.6
pytz
redis==4.3.1
fuzzywuzzy
openpyxl
openpyxl==3.0.9
matplotlib==3.7.1
matplotlib-inline==0.1.3
Markdown==3.3.7
APScheduler==3.10.4
python-Levenshtein